% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelPitching.R
\name{combine_pitcher_stints}
\alias{combine_pitcher_stints}
\title{Combine pitcher stints}
\usage{
combine_pitcher_stints(data)
}
\arguments{
\item{data}{A data frame containing pitching stats. Should be similar to Pitching from the Lahman package.}
}
\value{
A data frame conating pitching stats aggregated over all stints for a given player and season.
}
\description{
Combine pitching stats over all the stints for a given player and season.
}
\seealso{
\code{\link{get_pitching_stats}}
}
