% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelTeams.R
\name{get_games_for_year}
\alias{get_games_for_year}
\title{Get the expected number of games for a season.}
\usage{
get_games_for_year(year)
}
\arguments{
\item{year}{integer}
}
\description{
Get the expected number of games for a season. This function does a very
basic switch-case to distinguish 3 eras; pre-1902 (140 game), 
pre-expansion (154 games), and present day (162 games).
}
\seealso{
\code{\link{get_pa_for_year}} \code{\link{get_ipouts_for_year}}
}
