% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelTeams.R
\name{get_ipouts_for_year}
\alias{get_ipouts_for_year}
\title{Get the typical IPouts for a team for a year.}
\usage{
get_ipouts_for_year(year, ipouts_per_game = 27)
}
\arguments{
\item{year}{integer}

\item{ipouts_per_game}{numeric
The average number of PA per game. Defaults to 27.}
}
\description{
Get the typical IPouts for a team for a year.
}
\seealso{
\code{\link{get_pa_for_year}} \code{\link{get_game_for_year}}
}
