% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelTeams.R
\name{get_pa_for_year}
\alias{get_pa_for_year}
\title{Get the typical PA for a team for a season.}
\usage{
get_pa_for_year(year, pa_per_game = 38)
}
\arguments{
\item{year}{integer}

\item{pa_per_game}{numeric
The average number of PA per game. Defaults to 38.}
}
\description{
Get the typical PA for a team for a season.
}
\seealso{
\code{\link{get_ipouts_for_year}} \code{\link{get_game_for_year}}
}
