% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_primary_pos}
\alias{get_primary_pos}
\title{Get primary position}
\usage{
get_primary_pos(year = NULL)
}
\value{
A data frame containing playerID, yearID, primary position, 
and number of games at primary position.
}
\description{
Gets primary position based on `Fielding` table from the 
`Lahman` package. Uses the categories from the Lahman Fielding table 
so all OF positons are collapsed into `OF`. Main use is to filter
pitchers before computing seasonal averages batting stats.
}
\examples{
primary_pos <- get_primary_pos()
primary_pos \%>\% group_by(POS) \%>\% summarise(n=n())
}
