% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelBatting.R
\name{get_seasonal_averages_batting}
\alias{get_seasonal_averages_batting}
\title{Get seasonal averages for batting stats}
\usage{
get_seasonal_averages_batting(data)
}
\arguments{
\item{data}{A data frame containing batting stats. Should be similar to 
Lahman::Batting}
}
\value{
A data frame containing league averages for batting stats
}
\description{
Get seasonal averages for batting stats
}
