% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelPitching.R
\name{get_seasonal_averages_pitching}
\alias{get_seasonal_averages_pitching}
\title{Get seasonal averages for pitching stats}
\usage{
get_seasonal_averages_pitching(data)
}
\arguments{
\item{data}{A data frame containing pitching stats. Should be similar to 
Lahman::Pitching}
}
\value{
A data frame containing league averages for pitching stats
}
\description{
Get seasonal averages for pitching stats
}
