% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelTeams.R
\name{get_team_projected_pitching}
\alias{get_team_projected_pitching}
\title{Get team projected pitching}
\usage{
get_team_projected_pitching(projected_pitching)
}
\value{
A data frame with projected totals aggregated at the team-season level. 
Roster is taken from players that began the *following year* with the team.
}
\description{
Get team projected pitching
}
