% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelTeams.R
\name{get_team_projected_wins}
\alias{get_team_projected_wins}
\title{Get team projected wins}
\usage{
get_team_projected_wins(team_projected_batting = NULL,
  marcels_batting = NULL, roster_batting = NULL,
  team_projected_pitching = NULL, marcels_pitching = NULL,
  roster_pitching = NULL, team_mapping = NULL)
}
\value{
A data frame with projected win total aggregated at the team-season level.
}
\description{
Get team projected wins
}
\seealso{
\code{\link{get_team_projected_pitching}, \link{get_team_projected_batting}}
}
