\name{marcelR-package}
\alias{marcelR-package}
\alias{marcelR}
\docType{package}
\title{
\packageTitle{marcelR}
}
\description{
\packageDescription{marcelR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{marcelR}
\packageIndices{marcelR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{marcelR}

Maintainer: \packageMaintainer{marcelR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
