% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcels.R
\docType{data}
\name{marcels}
\alias{marcels}
\title{Marcel projections for batters and pitchers.}
\format{An object of class \code{list} of length 3.}
\usage{
marcels
}
\description{
A dataset containing marcel projections for batters & pitchers.
}
\examples{
marcels$Batting
}
\keyword{datasets}
