% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelPitching.R
\name{apply_marcel_pitching}
\alias{apply_marcel_pitching}
\title{Apply marcels for batters}
\usage{
apply_marcel_pitching(data, metric, age_adjustment_fun, metric_weights = c(3,
  2, 1), playing_time_weights = c(0.5, 0.1, 0))
}
\arguments{
\item{data}{data.frame}

\item{metric}{character
name of the metric, e.g. HR}

\item{age_adjustment_fun}{A callable to make the age adjustment}

\item{metric_weights}{An array with the weights to give to the projected stats for the previous seasons. 
The ith elemnt is the weight for the season i years previous. The default is c(3, 2, 1).}

\item{playing_time_weights}{An array with the weights to be used for projecting playing time.}
}
\value{
A data frame containgMrcel projections.
}
\description{
Apply marcels for batters
}
\examples{
a <- get_pitching_stats() \%>\% filter(yearID>=2001, yearID<=2003) \%>\% filter(POS=="P")
b <- tbl_df(append_previous_years(a, 
                                  get_seasonal_averages_pitching, 
                                  previous_years = 3))
mcl = apply_marcel_batting(b, "R", age_adjustment_reciprocal)
mcl \%>\% filter(projectedYearID==2004, playerID=='martipe02')

}
\seealso{
\code{\link{apply_marcel_batting}}, \code{\link{get_team_projected_pitching}}
}
