% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelBatting.R
\name{get_batting_stats}
\alias{get_batting_stats}
\title{Get stats for batters}
\usage{
get_batting_stats(PrimaryPosition = NULL)
}
\arguments{
\item{PrimaryPosition}{A data frame containing primary position. If not provided, 
it will be generated by the function.}
}
\value{
A data frame containing batting stats, primary position, and age. Batting stats are aggregated over
all stints.
}
\description{
Gets stats for batters. Pulls data from Lahman:::battingStats(), 
aggregates over stints and appends primary position and age.
}
\seealso{
\code{\link{get_primary_pos}} \code{\link{combine_batter_stints}}
}
