% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelPitching.R
\name{get_pitching_stats}
\alias{get_pitching_stats}
\title{Get stats for pitchers}
\usage{
get_pitching_stats(PrimaryPosition = NULL)
}
\arguments{
\item{PrimaryPosition}{A data frame containing primary position. If not provided, 
will be generated by the function.}
}
\value{
A data frame containing pitching stats and primary position. pitching stats are aggregated over
all stints.
}
\description{
Gets stats for pitchers. Similar to `Lahman:::pitchingStats()`, 
but aggregates over stints and appends primary position.
}
\seealso{
\code{\link{get_primary_pos}} \code{\link{combine_pitcher_stints}}
}
