% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcelTeams.R
\name{get_team_projected_batting}
\alias{get_team_projected_batting}
\title{Get team projected batting}
\usage{
get_team_projected_batting(marcels_batting, roster)
}
\arguments{
\item{marcels_batting}{A data frame containing batting projections}

\item{roster}{A data frame containing roster. It must have playerID and teamID columns.}
}
\value{
A data frame with projected totals aggregated at the team-season level. 
Roster is taken from players that began the following year with the team.
}
\description{
Get team projected batting
}
